Move idf_tools test stub into a fixture module
